Move QA tasks for PIDMC to DPG
- Avoid creating output histograms if not needed
- Move McParticles_000 to McParticles
