+++ Common/TableProducer/PID/CMakeLists.txt
@@ -26,11 +26,6 @@ o2physics_add_dpl_workflow(pid-tof-beta
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore
                     COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(pid-tof-qa-mc
-                    SOURCES qaTOFMC.cxx
-                    PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore
-                    COMPONENT_NAME Analysis)
-
 o2physics_add_dpl_workflow(pid-tof-full
                     SOURCES pidTOFFull.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore
@@ -46,11 +41,6 @@ o2physics_add_dpl_workflow(pid-tpc
 o2physics_add_dpl_workflow(pid-tpc-full
                     SOURCES pidTPCFull.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore ONNXRuntime::ONNXRuntime
-                    COMPONENT_NAME Analysis)
-
-o2physics_add_dpl_workflow(pid-tpc-qa-mc
-                    SOURCES qaTPCMC.cxx
-                    PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore
                     COMPONENT_NAME Analysis)
 
 # HMPID
