+++ DPG/Tasks/CMakeLists.txt
@@ -36,8 +36,19 @@ o2physics_add_dpl_workflow(pid-tof-qa
                     SOURCES qaPIDTOF.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore
                     COMPONENT_NAME Analysis)
+
+o2physics_add_dpl_workflow(pid-tof-qa-mc
+                    SOURCES qaTOFMC.cxx
+                    PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore
+                    COMPONENT_NAME Analysis)
+
 ## TPC
 o2physics_add_dpl_workflow(pid-tpc-qa
                     SOURCES qaPIDTPC.cxx
+                    PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore
+                    COMPONENT_NAME Analysis)
+
+o2physics_add_dpl_workflow(pid-tpc-qa-mc
+                    SOURCES qaTPCMC.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore
                     COMPONENT_NAME Analysis)
